fix(agents): satisfy directive persistence gates
