+++ src/agents/embedded-agent-runner/run/payloads.delivery-recovery.test.ts
@@ -1,0 +1,44 @@
+import type { AssistantMessage } from "openclaw/plugin-sdk/llm";
+import { describe, expect, it } from "vitest";
+import { buildPayloads } from "./payloads.test-helpers.js";
+
+describe("buildEmbeddedRunPayloads delivery recovery", () => {
+  it("uses persisted delivery facts for a recovered final assistant", () => {
+    const payloads = buildPayloads({
+      lastAssistant: {
+        role: "assistant",
+        stopReason: "stop",
+        content: [{ type: "text", text: "Recovered answer" }],
+        openclawDelivery: {
+          audioAsVoice: true,
+          replyToCurrent: true,
+          replyToId: "message-7",
+        },
+      } as AssistantMessage,
+    });
+
+    expect(payloads).toEqual([
+      expect.objectContaining({
+        text: "Recovered answer",
+        audioAsVoice: true,
+        replyToCurrent: true,
+        replyToId: "message-7",
+      }),
+    ]);
+  });
+
+  it("does not recover delivery facts by parsing a pre-upgrade assistant", () => {
+    const payloads = buildPayloads({
+      lastAssistant: {
+        role: "assistant",
+        stopReason: "stop",
+        content: [{ type: "text", text: "[[reply_to:message-7]] Recovered answer" }],
+      } as AssistantMessage,
+    });
+
+    expect(payloads).toHaveLength(1);
+    expect(payloads[0]?.text).toBe("Recovered answer");
+    expect(payloads[0]).not.toHaveProperty("replyToCurrent");
+    expect(payloads[0]).not.toHaveProperty("replyToId");
+  });
+});
