+++ src/agents/sessions/agent-session-base.ts
@@ -15,6 +15,7 @@ import type {
   AgentSessionEventListener,
   AgentSessionWriteSettlementRunner,
 } from "./agent-session-types.js";
+import { replaceAgentMessageInPlace } from "./agent-session-utils.js";
 import { formatNoApiKeyFoundMessage } from "./auth-guidance.js";
 import {
   type ExtensionCommandContextActions,
@@ -421,7 +422,7 @@ export abstract class AgentSessionBase {
         if (event.message.role === "assistant") {
           const persisted = this.sessionManager.getEntry(entryId);
           if (persisted?.type === "message" && persisted.message.role === "assistant") {
-            this.replaceMessageInPlace(event.message, persisted.message);
+            replaceAgentMessageInPlace(event.message, persisted.message);
           }
           this.lastAssistantEntryId = entryId;
         } else if (event.message.role === "user") {
@@ -482,21 +483,6 @@ export abstract class AgentSessionBase {
     return undefined;
   }
 
-  private replaceMessageInPlace(target: AgentMessage, replacement: AgentMessage): void {
-    // Agent-core stores the finalized message object in its state before emitting message_end.
-    // SessionManager persistence happens later in handleAgentEvent() with event.message.
-    // Mutating this object in place keeps agent state, later turn/agent events, listeners,
-    // and the eventual SessionManager.appendMessage(event.message) persistence in sync.
-    if (target === replacement) {
-      return;
-    }
-
-    for (const key of Object.keys(target)) {
-      Reflect.deleteProperty(target, key);
-    }
-    Object.assign(target, replacement);
-  }
-
   /** Emit extension events based on agent events */
   private async emitExtensionEvent(event: AgentEvent): Promise<boolean> {
     if (event.type === "agent_start") {
@@ -540,7 +526,7 @@ export abstract class AgentSessionBase {
       };
       const replacement = await this.currentExtensionRunner.emitMessageEnd(extensionEvent);
       if (replacement) {
-        this.replaceMessageInPlace(event.message, replacement);
+        replaceAgentMessageInPlace(event.message, replacement);
         return true;
       }
     } else if (event.type === "tool_execution_start") {
