+++ src/agents/sessions/session-manager-entries.ts
@@ -2,7 +2,7 @@ import {
   readActiveTranscriptEntryAnchor,
   type TranscriptEntryAnchor,
 } from "../../config/sessions/session-accessor.js";
-import { applyAssistantDeliveryDirectives } from "../../config/sessions/transcript-assistant-message.js";
+import { applyAssistantDeliveryDirectives } from "../../config/sessions/transcript-assistant-delivery.js";
 import { isSessionTranscriptSideAppendEntry } from "../../config/sessions/transcript-tree.js";
 import type { ImageContent, Message, TextContent } from "../../llm/types.js";
 import {
@@ -136,7 +136,9 @@ export class SessionManagerEntries extends SessionManagerPersistence {
     message: Message | CustomMessage | BashExecutionMessage,
     options?: AppendPersistenceOptions,
   ): { entryId: string; anchor?: TranscriptEntryAnchor } {
-    if (message.role === "assistant") applyAssistantDeliveryDirectives(message);
+    if (message.role === "assistant") {
+      applyAssistantDeliveryDirectives(message);
+    }
     if (options?.idempotencyLookup !== "caller-checked") {
       const currentUserId = this.resolveCurrentKeyedUserId(message);
       if (currentUserId) {
